requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
google-generativeai>=0.3.0
pyttsx3>=2.90
python-dotenv>=1.0.0 